+++ src/core/actions.rs
@@ -20,11 +20,11 @@ use crate::{
 
 use log::{debug, error};
 
-use super::error::ScorchError;
+use anyhow::{anyhow, Context, Result};
 
 ///URL to download
 ///file name to save to
-pub async fn download_file(url: &str, file_path: PathBuf) -> Result<File, ScorchError> {
+pub async fn download_file(url: &str, file_path: PathBuf) -> Result<File> {
     let client = Client::new();
 
     //send the request
@@ -32,15 +32,15 @@ pub async fn download_file(url: &str, file_path: PathBuf) -> Result<File, Scorch
         .get(url)
         .send()
         .await
-        .map_err(|_| (format!("Unable to GET from {}", url)))?;
+        .context(format!("Unable to GET from {}", url))?;
 
     if !res.status().is_success() {
         error!("Got bad response from thunderstore");
         error!("{:?}", res);
-        return Err(format!("{} at URL {}", res.status(), url).into());
+        return Err(anyhow!("{} at URL {}", res.status(), url));
     }
 
-    let file_size = res.content_length().ok_or(format!(
+    let file_size = res.content_length().ok_or(anyhow!(
         "Unable to read content length of response from {}",
         url
     ))?;
@@ -58,18 +58,14 @@ pub async fn download_file(url: &str, file_path: PathBuf) -> Result<File, Scorch
     let mut stream = res.bytes_stream();
     debug!("Starting download from {}", url);
     while let Some(item) = stream.next().await {
-        let chunk = item.map_err(|_| ("Error downloading file :(".to_string()))?;
-        file.write_all(&chunk).map_err(|e| {
-            error!("Error while writing download chunk to file");
-            error!("{}", e);
-            format!("Error writing to file {}", file_path.display())
-        })?;
+        let chunk = item.context("Error downloading file :(")?;
+        file.write_all(&chunk)
+            .context("Failed to write chunk to file")?;
         let new = min(downloaded + (chunk.len() as u64), file_size);
         downloaded = new;
         pb.set_position(new);
     }
-    let finished = File::open(&file_path)
-        .map_err(|_| (format!("Unable to open finished file {}", file_path.display())))?;
+    let finished = File::open(&file_path).context("Unable to open finished file")?;
     debug!("Finished download to {}", file_path.display());
 
     pb.finish_with_message(format!(
@@ -79,20 +75,20 @@ pub async fn download_file(url: &str, file_path: PathBuf) -> Result<File, Scorch
     Ok(finished)
 }
 
-pub fn uninstall(mods: Vec<&PathBuf>) -> Result<(), String> {
-    mods.iter().for_each(|p| {
-        fs::remove_dir_all(p).expect("Unable to remove directory");
+pub fn uninstall(mods: Vec<&PathBuf>) -> Result<()> {
+    for p in mods {
+        fs::remove_dir_all(p).context("Unable to remove directory")?;
         println!("Removed {}", p.display());
-    });
+    }
     Ok(())
 }
 
-pub fn install_mod(zip_file: &File, config: &Config) -> Result<InstalledMod, String> {
+pub fn install_mod(zip_file: &File, config: &Config) -> Result<InstalledMod> {
     debug!("Starting mod insall");
     let mods_dir = config
         .mod_dir()
         .canonicalize()
-        .map_err(|_| "Couldn't resolve mods directory path".to_string())?;
+        .context("Couldn't resolve mods directory path")?;
     //Get the package manifest
     let mut manifest = String::new();
     //Extract mod to a temp directory so that we can easily see any sub-mods
@@ -105,20 +101,15 @@ pub fn install_mod(zip_file: &File, config: &Config) -> Result<InstalledMod, Str
             .to_string(),
     );
     {
-        let mut archive =
-            ZipArchive::new(zip_file).map_err(|_| ("Unable to read zip archive".to_string()))?;
+        let mut archive = ZipArchive::new(zip_file).context("Unable to read zip archive")?;
 
         archive
             .by_name("manifest.json")
-            .map_err(|_| "Couldn't find manifest file".to_string())?
+            .context("Couldn't find manifest file")?
             .read_to_string(&mut manifest)
             .unwrap();
 
-        fs::create_dir_all(&temp_dir).map_err(|e| {
-            error!("Unable to create temp directory");
-            error!("{}", e);
-            "Unable to create temp directory".to_string()
-        })?;
+        fs::create_dir_all(&temp_dir).context("Unable to create temp directory")?;
         for i in 0..archive.len() {
             let mut file = archive.by_index(i).unwrap();
             let out = temp_dir.join(&file.enclosed_name().unwrap());
@@ -132,18 +123,17 @@ pub fn install_mod(zip_file: &File, config: &Config) -> Result<InstalledMod, Str
 
             debug!("Extracting file to {}", out.display());
             if (*file.name()).ends_with('/') {
-                fs::create_dir_all(&out).unwrap();
+                fs::create_dir_all(&out)?;
                 continue;
             } else if let Some(p) = out.parent() {
-                fs::create_dir_all(&p).unwrap();
+                fs::create_dir_all(&p)?;
             }
             let mut outfile = OpenOptions::new()
                 .create(true)
                 .write(true)
                 .truncate(true)
-                .open(&out)
-                .unwrap();
-            io::copy(&mut file, &mut outfile).unwrap();
+                .open(&out)?;
+            io::copy(&mut file, &mut outfile).context("Unable to copy file")?;
         }
     }
     let mut mods = vec![];
@@ -165,31 +155,26 @@ pub fn install_mod(zip_file: &File, config: &Config) -> Result<InstalledMod, Str
     }
 
     if mods.is_empty() {
-        error!("Didn't find any directories in extracted archive");
-        return Err("Couldn't find a directory to copy".to_string());
+        return Err(anyhow!("Couldn't find a directory to copy"));
     }
 
-    mods.iter()
+    for p in mods
+        .iter()
         .map(|p| (&p.path, mods_dir.join(p.path.file_name().unwrap())))
-        .for_each(|p| {
-            if p.1.exists() {
-                fs::remove_dir_all(&p.1).unwrap();
-            }
-            fs::rename(p.0, p.1).unwrap();
-        });
+    {
+        if p.1.exists() {
+            fs::remove_dir_all(&p.1)?;
+        }
+        fs::rename(p.0, p.1)?;
+    }
 
     for mut m in mods.iter_mut() {
         m.path = mods_dir.join(m.path.file_name().unwrap());
     }
 
-    let manifest: Manifest =
-        serde_json::from_str(&manifest).map_err(|_| "Unable to parse manifest".to_string())?;
+    let manifest: Manifest = serde_json::from_str(&manifest).context("Unable to parse manifest")?;
 
-    fs::remove_dir_all(&temp_dir).map_err(|e| {
-        error!("Unable to remove temp directory");
-        error!("{}", e);
-        "Unable to remove temp directory".to_string()
-    })?;
+    fs::remove_dir_all(&temp_dir).context("Unable to remove temp directory")?;
 
     Ok(InstalledMod {
         package_name: manifest.name,