+++ src/core/error.rs
@@ -1,34 +1,35 @@
 use std::{error::Error, fmt::Display};
 
 #[derive(Debug)]
-pub(crate) struct ScorchError(pub String);
+pub(crate) enum ScorchError {
+    FsError(Box<dyn Error>),
+    NotFound(String),
+    NetworkError(Box<dyn Error>),
+    Generic(Box<dyn Error>),
+}
 
 impl Display for ScorchError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
+        match self {
+            e => write!(f, "{}", e),
+        }
     }
 }
 
-impl From<&dyn Error> for ScorchError {
-    fn from(e: &dyn Error) -> Self {
-        ScorchError(e.to_string())
-    }
-}
+// impl From<String> for Error {
+//     fn from(e: String) -> Self {
 
-impl From<String> for ScorchError {
-    fn from(e: String) -> Self {
-        ScorchError(e)
-    }
-}
+//     }
+// }
 
-impl From<&str> for ScorchError {
-    fn from(e: &str) -> Self {
-        ScorchError(e.to_string())
-    }
-}
+// impl From<&str> for Error {
+//     fn from(e: &str) -> Self {
+//         Error(e.to_string())
+//     }
+// }
 
 impl From<std::io::Error> for ScorchError {
     fn from(e: std::io::Error) -> Self {
-        ScorchError(e.to_string())
+        ScorchError::FsError(Box::new(e))
     }
 }